fix: setTouched on Select Close instead of Open
This will set `touched` on closing the option instead of open.

A typical HTML Select in Angular will set touched once the user
mouse clicks out or tabs (keyboard) out. This will emulate that
same behaviour. Setting `touched` right away on open causes
weird user experience when displaying error message that 
typically depend on `touched` being set. We want to let
the user interact with the select and once done interaction,
it is to be set `touched`.