Prevent clear list when pass multiple
I pass multiple arugment value is false explicit as <ng-select [multiple]="multiple"> and sometime when I navigate to page with ng-select list is empty. If I don't pass and multiple is default false it work fine. By default and pass multiple = false expected will same behavior but it's not. Try to fixed and seems it work well