feat: ignore ctrl+enter like a normal select
The rationale of this change is as follows:

Having an input text focused, one can submit a form by pressing enter.
But this does not work when other elements like textarea and select are
in focus.

To circumvent this, one can define ctrl+enter to submit a form,
regardless which element is in focus.

Prior to this change, when pressing ctrl+enter, the user would see the
ng-select menu reopen whilst submitting which is confusing.

By ignoring the control key press, ng-select does not reopen and behaves
like a standard select.