fix: add themes directory to package.json exports
This PR enables importing files from the "themes" directory when using Yarn with PnP. When used, the "exports" field in a package.json is meant to conceal all files not specified in "exports". Until now the exclusion of the "themes" directory wasn't an apparent issue because no other package manager/strategy is able to enforce disallowing importing files not specified in the "exports" field but Yarn with PnP enforces it.

fixes #2065