fix(docs): The issues with the application build and deployment have been addressed.
The changes made include the import of APP_BASE_HREF in 'app.module.ts' which was commented out previously. In the 'release.yml' file, the path for deployment has been changed to 'dist/demo' instead of 'dist/demo/browser'. This is also reflected in 'package.json' where the path for 'copy-examples' has been updated.
The 'angular.json' file now uses '@angular-devkit/build-angular:browser-esbuild' as the builder and adds 'deployUrl' to the options. It is believed these changes will fix the build issues observed previously.